daemon: Allow host fw w/o the remote-node identity
In 9654cb4, the Cilium agent was made to fatal if the host firewall was
enabled while the remote-node identity was disabled. This was based on
tests using existing policy examples during which the connectivity check
was failing.

It turns out the host firewall works fine without the remote-node
identity. The policy just needs to be adapted to take that into account,
by using the node CIDR in place of the remote-node identity. This commit
thus reverts 9654cb4.

Fixes: 9654cb4 ("daemon: Fatal if host fw is enabled w/o remote node identities")
Signed-off-by: Paul Chaignon <paul@cilium.io>